+++ src/backend/patterns/BonusHint.java
@@ -1,0 +1,5 @@
+package backend.patterns;
+
+public interface BonusHint {
+    String getHint(int guess, int target);
+}
