+++ src/backend/patterns/HighLowHintDecorator.java
@@ -1,0 +1,16 @@
+package backend.patterns;
+
+public class HighLowHintDecorator implements BonusHint {
+    private BonusHint bonusHint;
+
+    public HighLowHintDecorator(BonusHint bonusHint) {
+        this.bonusHint = bonusHint;
+    }
+
+    @Override
+    public String getHint(int guess, int target) {
+        String baseHint = bonusHint.getHint(guess, target);
+        String highLowHint = guess > target ? "Too high!" : "Too low!";
+        return baseHint + " " + highLowHint;
+    }
+}
